backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
firebase-admin>=6.0.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
Pillow>=9.0.0
